+++ scripts/ci/pylint_todo.txt
@@ -58,7 +58,7 @@
 ./airflow/providers/amazon/aws/operators/sagemaker_base.py
 ./airflow/providers/amazon/aws/operators/sagemaker_endpoint_config.py
 ./airflow/providers/amazon/aws/operators/sagemaker_endpoint.py
-./airflow/contrib/operators/sagemaker_model_operator.py
+./airflow/providers/amazon/aws/operators/sagemaker_model.py
 ./airflow/contrib/operators/sagemaker_training_operator.py
 ./airflow/contrib/operators/sagemaker_transform_operator.py
 ./airflow/contrib/operators/sagemaker_tuning_operator.py
